requirements.txt
# Placeholder
torch>=2.0.0
opencv-python
ultralytics
numpy
PyYAML
matplotlib
scikit-learn
streamlit
tensorboard
